Build group relationships
Working from https://github.com/creativice/graph-airwatch/pull/2/files,
build group relationships in current codebase. Note that this relates
direct children since each group is processed and should produce a
complete tree. It may be an optimization to only fetch children of a
group that is not a child of another group, since it appears that the
complete heirarchy is included in the response for a root group.